Make the new removal logic its own command instead of replacing
