+++ addon/utils/ce/raw-editor.ts
@@ -61,9 +61,10 @@ import {
   MarkSpec,
 } from '@lblod/ember-rdfa-editor/model/mark';
 import AddMarkToRangeCommand from '@lblod/ember-rdfa-editor/commands/add-mark-to-range-command';
-import RemoveMarkFromRangeCommand from '@lblod/ember-rdfa-editor/commands/remove-mark-from-range-command';
+import RemoveMarksFromRangesCommand from '@lblod/ember-rdfa-editor/commands/remove-marks-from-ranges-command';
 import RemoveMarkCommand from '@lblod/ember-rdfa-editor/commands/remove-mark-command';
 import MatchTextCommand from '@lblod/ember-rdfa-editor/commands/match-text-command';
+import RemoveMarkFromRangeCommand from '@lblod/ember-rdfa-editor/commands/remove-mark-from-range-command';
 
 export interface RawEditorProperties {
   baseIRI: string;
@@ -189,6 +190,7 @@ export default class RawEditor {
     this.registerCommand(new SetPropertyCommand(this.model));
     this.registerCommand(new AddMarkToRangeCommand(this.model));
     this.registerCommand(new RemoveMarkFromRangeCommand(this.model));
+    this.registerCommand(new RemoveMarksFromRangesCommand(this.model));
     this.registerCommand(new RemoveMarkCommand(this.model));
     this.registerCommand(new MatchTextCommand(this.model));
     this.registerMark(highlightMarkSpec);
